+++ src/routes.js
@@ -1,13 +1,13 @@
 const { Router } = require('express');
 const BusController = require('./controllers/BusController');
-
+const StoredController = require('./controllers/StoredController');
 
 const routes = Router();
 
-routes.post('/bus/InsertBus', BusController.saveBus);
-routes.post('/bus/InsertHorarios', BusController.saveHorario);
-routes.post('/bus/InsertPonto', BusController.savePonto);
-routes.post('/bus/InsertParadas',BusController.saveParada);
+routes.post('/bus/InsertBus', StoredController.saveBus);
+routes.post('/bus/InsertHorarios', StoredController.saveHorario);
+routes.post('/bus/InsertPonto', StoredController.savePonto);
+routes.post('/bus/InsertParadas',StoredController.saveParada);
 
 routes.get('/bus/BuscarLinhas', BusController.GetLinhas);
 routes.get('/bus/BuscarLinhasByNum', BusController.GetLinhaByNum);
